requirement.txt
boto
awscli
pypdf
langchain
streamlit
faiss-cpu